+++ test/valueFormatterTest.ts
@@ -40,6 +40,50 @@ import { valueFormatter } from "./../src/valueFormatter";
 
 describe("ValueFormatter", () => {
 
+    describe("checkValueInBounds", () => {
+        it("target is greater than bounds", () => {
+            const res = valueFormatter.checkValueInBounds(5, -4, 4);
+            expect(res).toBe(4);
+        });
+
+        it("target inside bounds", () => {
+            const res = valueFormatter.checkValueInBounds(3, 0, 14);
+            expect(res).toBe(3);
+        });
+
+        it("target is lesser that bounds", () => {
+            const res = valueFormatter.checkValueInBounds(-3, 0, 14);
+            expect(res).toBe(0);
+        });
+
+        it("target is lesser that bounds with lessMinReplacement as null", () => {
+            const res = valueFormatter.checkValueInBounds(-3, 0, 14, null);
+            expect(res).toBeNull();
+        });
+
+        it("max is undefined", () => {
+            let res = valueFormatter.checkValueInBounds(5, -4, null);
+            expect(res).toBe(5);
+
+            res = valueFormatter.checkValueInBounds(5, -4, undefined);
+            expect(res).toBe(5);
+
+            res = valueFormatter.checkValueInBounds(-5, -4, null, null);
+            expect(res).toBeNull();
+        });
+
+        it("min is undefined", () => {
+            let res = valueFormatter.checkValueInBounds(5, null, 15);
+            expect(res).toBe(5);
+
+            res = valueFormatter.checkValueInBounds(5, undefined, 15);
+            expect(res).toBe(5);
+
+            res = valueFormatter.checkValueInBounds(-2, null, -3);
+            expect(res).toBe(-3);
+        });
+    });
+
     describe("format", () => {
         it("format null", () => {
             expect(valueFormatter.format(null)).toBe("(Blank)");