+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 4.1.3
+* Was added checkValueInBounds function for formatter.
+
 # 4.1.2
 * Changed mmm, mmmm and MMMM flag to use the month index value (M) instead of minutes (m) to return correct month names.